requirements.txt
Flask==2.0.3
gunicorn==20.1.0
psycopg2-binary==2.9.3
Werkzeug==2.0.3
Flasgger==0.9.5
flask-cors==3.0.10
pandas==1.3.5  # ✅ pandas 추가
numpy==1.21.6 # pandas 1.3.5 버전과 호환되는 버전
xlsxwriter==3.0.8  # ✅ Excel 생성 라이브러리 추가 (Excel 다운로드 기능을 위해 필요)